src/requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.28
psycopg2-binary==2.9.9
redis==5.0.1
drf-yasg==1.21.7
gunicorn==21.2.0
django-modeltranslation==0.18.11
channels==4.0.0
daphne==4.0.0
channels_redis==4.2.0
Pillow==10.2.0
requests==2.28.1
django-modeltranslation==0.18.11
